Fork update

+++ src/main/java/es/tid/bgp/bgp4/update/tlv/linkstate_attribute_tlvs/DefaultTEMetricLinkAttribTLV.java
@@ -68,7 +68,7 @@ public class DefaultTEMetricLinkAttribTLV extends BGP4TLVFormat {
 			}	
 	}
 
-	public float getLinkMetric() {
+	public long getLinkMetric() {
 		return linkMetric;
 	}
